src/main/webapp/app/entities/condicoes/route/condicoes-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { ICondicoes, Condicoes } from '../condicoes.model';
import { CondicoesService } from '../service/condicoes.service';

import { CondicoesRoutingResolveService } from './condicoes-routing-resolve.service';

describe('Condicoes routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: CondicoesRoutingResolveService;
  let service: CondicoesService;
  let resultCondicoes: ICondicoes | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    routingResolveService = TestBed.inject(CondicoesRoutingResolveService);
    service = TestBed.inject(CondicoesService);
    resultCondicoes = undefined;
  });

  describe('resolve', () => {
    it('should return ICondicoes returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultCondicoes = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultCondicoes).toEqual({ id: 123 });
    });

    it('should return new ICondicoes if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultCondicoes = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultCondicoes).toEqual(new Condicoes());
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as Condicoes })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultCondicoes = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultCondicoes).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
